Update Readme.md
This project demonstrates the Database connectivity in Java It has 2 files.

CountryDemo.java
DBUtil.java
DBUtil.java
This file manages the connection with SQL and returns the connection
CountryDemo.java
This file extends the DBUtil file. The Main method takes a manual SQL query string to extract every country from the SQL table. It has try and catch function to handle the connection exception. It finally prints out the country and total number of country present in the Database